Re-organizing generated file and providing entry point for Win and Doc
